+++ Assets/Scripts/Screamer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c10a677adb234131a46b80ae56490ddc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
